requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
joblib>=1.2.0
flask
flask-cors
gunicorn
psycopg2-binary
transformers>=4.30.0
torch>=2.0.0
